requirements.txt
httpx>=0.27
numpy>=1.26
openai-whisper>=20231117
psutil>=5.9
pyttsx3>=2.90
sounddevice>=0.4
opencv-python>=4.10
mss>=9.0
pystray>=0.19
Pillow>=10.4
torch>=2.2; platform_system=="Windows"
torchaudio>=2.2; platform_system=="Windows"
